project1/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-prometheus==2.4.1
googleapis-common-protos==1.70.0
grpcio==1.74.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
opentelemetry-api==1.36.0
opentelemetry-distro==0.57b0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-django==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
prometheus_client==0.22.1
protobuf==6.31.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
zipp==3.23.0
